sentiment_analytics/negapoji_analytics/requirements.txt
torch==1.4.0
torchtext==0.5.0
torchvision==0.5.0

janome
matplotlib
mecab-python3
mojimoji
numpy
spacy
tqdm
